Rework resource initial data upload.
The test_cl_get_info test before this change runs into out-of-memory.

As an optimization for single-device cases, we go ahead and allocate
a backing D3D12TranslationLayer (and D3D12 resource, either committed
or suballocated) at CL resource mem object creation time.

We also go ahead and upload initial data at this time. Since resource
creation is not synchronized with the device's worker thread, we used
the InitialData upload scenario. That means that temp memory allocated
for this op is not counted towards the current command list's memory usage,
meaning work is never opportunistically flushed if the only thing being
done is initial data uploads.

However:
The InitialData scenario was intended to be used only for the creation
of the CPrepareUpdateSubresources structure, which can safely be
unsynchronized with the rest of the immediate context, but not for the
FinalizeUpdateSubresources function, which must by synchronized with
the immediate context (for CLOn12, must be synchronized with the device's
task recording thread). This is a bug, but we just hadn't run content
which hit this race.

This change does two things:
1. Remove the optimization for creating translation layer resources early.
   This is sufficient to fix the test_cl_get_info test.
2. Move initial data upload to a task, which is submitted as a dependency
   of whatever task caused the resource to actually be created. This would
   also be sufficient to fix the test, as it would allow work to be
   opportunistically flushed by just initial data uploads.