Rework local size computation
See https://github.com/microsoft/OpenCLOn12/issues/50

The previous algorithm did not gracefully handle odd/prime sizes. It
would start with a default size, shrink to 1, and never expand again.

The new algorithm attempts to prime factorize the global size and add
that into the local size. This continues until it hits a sweet spot for
the device's wave size, >= min and < max (unless max == min). Failing
to find a sweet spot, it'll just attempt to use one of the global
dimensions as the local dimension.